+++ test/e2e/qa-lab/media/vision-routing-core.e2e.test.ts
@@ -28,6 +28,7 @@ const preparedVisionCatalog = vi.hoisted(() => [
 ]);
 
 vi.mock("../../../../src/agents/prepared-model-catalog.js", () => ({
+  loadProviderScopedThinkingCatalog: vi.fn(async () => []),
   loadPreparedModelCatalog: vi.fn(async () => preparedVisionCatalog),
 }));
